wrap up param generation
