+++ src/shared/parameter/BUILD
@@ -54,9 +54,9 @@ cc_library(
 )
 
 cc_test(
-    name = "cpp_config_test",
+    name = "config_test",
     srcs = [
-        "cpp_config_test.cpp",
+        "config_test.cpp",
     ],
     data = [
         "//shared/parameter/config_definitions:parameter_configs",
